--- a/src/materials/ComputeISoilStress.C
+++ b/src/materials/ComputeISoilStress.C
@@ -51,7 +51,9 @@ validParams<ComputeISoilStress>()
   params.addParam<std::vector<Real>>("p_ref",
                                      "The reference pressure at which "
                                      "the parameters are defined for "
-                                     "each soil layer.");
+                                     "each soil layer. If 'soil_type = "
+                                     "darendeli', then the reference "
+                                     "pressure must be input in kilopascals.");
   params.addParam<Real>("a0",
                         1.0,
                         "The first coefficient for pressure dependent yield strength "
@@ -205,6 +207,12 @@ ComputeISoilStress::ComputeISoilStress(const InputParameters & parameters)
                  ". Pressure dependency is set to true but a0, a1 and a2 are "
                  "set to 1.0, 0.0 and 0.0, respectively. Strength "
                  "pressure dependency is NOT turned on.");
+  if (_pressure_dependency && (_a0 == 0.0 && _a1 == 0.0 && _a2 == 0.0))
+    mooseError("Error in " + name() +
+                 ". When pressure dependency is turned on, "
+                 "all three strength coefficients, a0, a1, and a2, "
+                 "cannot simultaneously be set to 0.0. This "
+                 "combination results in division by 0.");
   if (_pressure_dependency && _p_ref.size() != _layer_ids.size())
     mooseError("Error in " + name() + ". When pressure dependency is turned on, "
                                       "a positive reference pressure "