--- a/test/tests/materials/isoil/HYS_stiffness_and_strength_pressure_dependency.i
+++ b/test/tests/materials/isoil/HYS_stiffness_and_strength_pressure_dependency.i
@@ -332,6 +332,7 @@
     type = PiecewiseLinear
     data_file = Displacement2.csv
     format = columns
+    scale_factor = 10000
   [../]
 []