--- a/test/tests/materials/isoil/HYS_stiffness_and_strength_pressure_dependency_noaction.i
+++ b/test/tests/materials/isoil/HYS_stiffness_and_strength_pressure_dependency_noaction.i
@@ -335,6 +335,7 @@
     type = PiecewiseLinear
     data_file = Displacement2.csv
     format = columns
+    scale_factor = 10000
   [../]
 []