--- a/test/tests/materials/isoil/HYS_stiffness_and_strength_pressure_dependency_noaction_AD.i
+++ b/test/tests/materials/isoil/HYS_stiffness_and_strength_pressure_dependency_noaction_AD.i
@@ -336,6 +336,7 @@
     type = PiecewiseLinear
     data_file = Displacement2.csv
     format = columns
+    scale_factor = 10000
   [../]
 []