… test/tests/materials/isoil/gold/HYS_stiffness_and_strength_pressure_dependency_out.e