MASTODON's ISoil module has been profiled using Google's performance tools. This has led to an understanding of which parts of the code of ISoil are having bottlenecks. Efforts were made to curtail such bottlenecks in ComputeISoilStress.C and ADComputeISoilStress.C. The following tasks were accomplished:
- [x] Addressed repeated but same calculations within the for loops in the above-mentioned .C files 
- [x] Addressed creation of variables multiple time within the for loops in the above-mentioned .C files
- [x] Fixed the tests file for ISoil to compare quantities having values of practical significance. Previously, values of the order ~10^-8 were being compared which practically means comparing zeros.

With the above adjustments, ISoil's computational efficiency has improved 2-2.5 times depending upon the problem. Please see below in the comments for details about the problem and the exact computational times before and after making improvements to ISoil.