@somu15 2019-09-23
### **Performance after removing bottle necks from ISoil**
Example 03 is the case study.

**With new code changes**
Without AD: 561.532s
With AD: 1477.213s

**Without new code changes**
Without AD: 685.801s
With AD: 2717.083s

Also, the attached plot compares the top soil acceleration with and without the new code changes (just in case). They match quite well.

![Comparison_09232019](https://user-images.githubusercontent.com/19966054/65455331-39a71f00-de04-11e9-8afb-ca7d3e4569b8.png)