@somu15 2019-10-18
### **3D Site response performance after removing bottle necks from ISoil**

**With new code changes**
Without AD: 7.83 hours
With AD: 10.42 hours

**Without new code changes**
Without AD: 8.02 hours
With AD: 18.43 hours