@sveerara 2019-10-18
@somu15, Thanks for updating the results from the 3D free-field analyses. It looks like AD does not help with the run times on larger problems as well. Do you see any improvement on the number of linear/nonlinear iterations between the no-AD and AD versions?