@somu15 2019-10-18
@sveerara Hi Swetha,

Attached are the plots of number of linear/nonlinear iterations vs time. It looks like no-AD takes less linear iterations, in general.

![Linear_Iterations](https://user-images.githubusercontent.com/19966054/67121349-86cea480-f1a8-11e9-83a8-4987ad380501.png)

![Nonlinear_Iterations](https://user-images.githubusercontent.com/19966054/67121358-8c2bef00-f1a8-11e9-8f04-c888496de9fc.png)