@sveerara 2019-10-18
I am surprised that linear iterations with AD is higher than that without AD. Is there a limit on the max. number of linear iterations in the Executioner block? Also, in the AD version, are the kernels using the displaced or the un-displaced mesh?