@sveerara 2019-10-18
l_max_its = 50 should be fine. 

I think using the displaced mesh for the inertial kernels might be causing the Jacobian to be incorrect, resulting in the higher number of linear iterations. The inertia kernels are not yet converted to the AD form. It might be worth re-running the AD version with all kernels in the undisplaced configuration.