@somu15 2019-10-18
@sveerara Ok, I will re-run the AD simulations in un-displaced configuration.

So, I should use `use_displaced_mesh = false` in the `GlobalParams` block and in the `DynamicTensorMechanics` and `Gravity` blocks. Correct?