@sveerara 2019-10-18
Thanks @somu15. You can either set use_displaced_mesh = false in the GlobalParams or in all the kernels - DynamicTensorMechanics, Gravity and InertialForce.