@somu15 2019-10-22
@sveerara Here are the new results.

**Linear iterations**
![Linear_Iterations](https://user-images.githubusercontent.com/19966054/67305275-52adf900-f4b2-11e9-8898-cf66fbbf478e.png)


**Nonlinear iterations**
![Nonlinear_Iterations](https://user-images.githubusercontent.com/19966054/67305296-5b9eca80-f4b2-11e9-95a2-673ede2ad783.png)